requirements.txt
gfootball
openai
pyyaml
numpy
python-dotenv
